faker.js para a pasta utils
